requirements.txt
playwright
crawl4ai
google-generativeai
instagrapi
requests
python-dotenv
pillow
